Always include line break node
Apodimark was trying to be helpful by not providing a new line node in the tree
if it came at the end of a string. We need this new line node right now so you can still
create multi-line items, and soon to use the new line node as an indicator
that we need to create a new item.